+++ stories/components/counter/docs/examples/counter_in_pills.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import Pills from 'intergalactic/pills';
-import { Text } from 'intergalactic/typography';
+import Pills from '@semcore/pills';
+import { Text } from '@semcore/typography';
 
 const Demo = () => (
-  <Pills defaultValue='all'>
+  <Pills defaultValue='all' aria-label='Pills with counters'>
     <Pills.Item value='all'>
       <Pills.Item.Text>All</Pills.Item.Text>
       <Pills.Item.Addon>