+++ semcore/feedback-form/CHANGELOG.md
@@ -2,13 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [6.37.1] - 2024-11-29
+## [6.37.1] - 2024-12-06
 
 ### Changed
 
-- FeedbackRating form a11y
+- FeedbackRating form a11y.
 
-## [6.37.0] - 2024-12-02
+### Fixed
+
+- There was always `aria-errormessage` in FeedbackItem, even if there was no message popup.
+
+## [6.37.0] - 2024-11-29
 
 ### Changed